declutter common.py: extract helpers and KVRouterProcess into their own modules
