+++ tests/router/test_router_e2e_with_trtllm.py
@@ -12,13 +12,12 @@ from typing import Any, Dict, Optional
 
 import pytest
 
-from tests.router.common import (  # utilities
+from tests.router.common import (
     _test_router_basic,
     _test_router_decisions,
     _test_router_indexers_sync,
-    generate_random_suffix,
-    get_runtime,
 )
+from tests.router.helper import generate_random_suffix, get_runtime
 from tests.utils.constants import DefaultPort
 from tests.utils.managed_process import ManagedProcess
 from tests.utils.port_utils import allocate_ports, deallocate_ports
